+++ NC-TASK/src/Task3/Building.java
@@ -8,7 +8,7 @@ public interface Building {
 
     int getAllSpace();
 
-    int getAllArea();
+    double getAllArea();
 
     int getAllRoom();
